Reorganize source tree
firebuild/ - supervisor executable sources
fbintercept/ - interceptor shared library sources
common/ - common files

Fixes #11.
