+++ .gitignore
@@ -12,6 +12,6 @@ Doxyfile
 Makefile
 build/
 cmake_install.cmake
-src/firebuild/fb-messages.pb.cc
-src/firebuild/fb-messages.pb.h
+fb-messages.pb.cc
+fb-messages.pb.h
 src/firebuild/firebuild
