+++ src/common/firebuild_common.cc
@@ -2,7 +2,7 @@
 /* This file is an unpublished work. All rights reserved. */
 
 
-#include "firebuild/firebuild_common.h"
+#include "common/firebuild_common.h"
 
 #include <netinet/in.h>
 #include <google/protobuf/message_lite.h>
