+++ src/common/firebuild_common.h
