+++ src/interceptor/env.c
@@ -2,7 +2,7 @@
 /* This file is an unpublished work. All rights reserved. */
 
 
-#include "firebuild/env.h"
+#include "interceptor/env.h"
 
 #include <unistd.h>
 #include <assert.h>
